toradex: integrate command whitelisting (upstream)
Integrate the command whitelisting feature which is part of the Secure
Boot hardening on U-Boot made by Toradex. The feature allows categories
of commands to be enabled/disabled depending on whether the device is
in open/closed state related to HAB/AHAB; the selection of allowed and
denied categories is taken from the control DTB having also hard-coded
sane defaults.

Upstream-Status: Inappropriate [TorizonCore specific]

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>