toradex: integrate command whitelisting (k3)
v1:

Integrate the command whitelisting feature which is part of the Secure
Boot hardening on U-Boot made by Toradex. The feature allows categories
of commands to be enabled/disabled depending on whether the device is
in open/closed state related to HAB/AHAB; the selection of allowed and
denied categories is taken from the control DTB having also hard-coded
sane defaults.

v2:

Add new exit code to indicate that a command failed to execute because
it was blocked by the whitelist feature. The numeric code chosen for
that purpose was 126, just like the code produced by Bash when it cannot
execute a given command.

Upstream-Status: Inappropriate [Torizon OS specific]

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>