toradex: verdin-imx8mm: Enable usb sdp support
Add necessary configuration options and device tree changes
for usb sdp support. USB driver model disabled for SPL to reduce
size as it is not needed for now for usb to works.

Upstream-Status: Submitted [https://lore.kernel.org/all/20230822221726.8355-1-marcel@ziswiler.com]

The changes are almost a backport from
commit 1c2567a3f129 ("verdin-imx8mm: enable usb sdp spl recovery support")
of u-boot-toradex next that was already submmited to upstream.
The differences to the patch are device tree syntax and getting
only the necessary configurations.

Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>