toradex: env: Enable env is nowhere for imx8 modules
The downstream code will not build when the environment on MMC is not
defined. Additionally, when the environment on EMMC is defined
simultaneously with ENV_IS_NOWHERE, it takes priority. This change makes
it possible to use ENV_IS_NOWHERE for recovery mode without the need
to enable the environment in MMC downstream.

Upstream-Status: Inappropriate [other]

This change is specific to the downstream code, as the solution in upstream
differs.

Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>