imx8: Enable env is nowhere for apalis and colibri
The downstream code will not build when the environment on MMC is not
defined. This change makes it possible to use ENV_IS_NOWHERE
for recovery mode without the need to enable the environment
in MMC downstream. This is the same correction of commit 8bb483b3cb3d
("toradex: env: Enable env is nowhere for imx8 modules") that solve it
for imx8mm/mp. The difference is that imx8 modules define the function
on another file.

Upstream-Status: Inappropriate [other]

This change is specific to the downstream code, as the solution in upstream
differs.

Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>