toradex: aquila-am69: fix dfu_alt_info_ram in env text
Following the move of the Aquila AM69 U-Boot environment to a text file in
commit d320af1a819d ("toradex: aquila-am69: move environment definition to
env file"), the dfu_alt_info_ram variable stopped functioning as expected.
This was due to the improper use of other variables like loadaddr,
scriptaddr, and ramdisk_addr_r which were not supported in the original
format.

This fix updates the dfu_alt_info_ram variable to use #define values for
scriptaddr and ramdisk_addr_r, while leveraging the macro CONFIG_SYS_LOAD_ADDR
for loadaddr.

Upstream-Status: Pending
Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>