aquila-am69: Fix combined firmware and related docs and env file
Commit 347fffe7 changed the offsets of the combined firmware
binary and the position of the U-Boot sector, but only for the
hs-fs binary. The documentation and U-Boot update wrapper
commands became out of sync with these changes.

Fix this by also updating the documentation and aquila-am69.env.
Apply the change to the gp and hs binaries as well.

Upstream-Status: Pending
Fixes: 347fffe72603 ("aquila-am69_*_defconfig: Adjust CONFIG_SYS_MMCSD_RAW_MODE_U_BOOT_SECTOR")
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>