verdin-imx8mp: Remove unnecessary pin configuration
The pin MX8MP_IOMUXC_NAND_READY_B__USDHC3_RESET_B is not used on the
Verdin SoM, and having it configured in u-boot breaks Cortex-M use
cases for that pin.

There's no need to send this patch upstream because this pin is already
not configured there.

Upstream-Status: Inappropriate [Other]
Related-to: ELB-6203
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>