board: toradex: smarc-imx95: notify Embedded Controller when firmware is loaded
The Embedded Controller provides a 'Firmware Loaded' command to signal
when the initial firmware has been loaded.

The SoC USB1 (3.0), normally connected to a USB 3.0 hub in host mode
(providing USB1, USB2, USB3, and USB4 SMARC host interfaces), is the only
interface that supports USB recovery. A mux allows switching SoC USB1 to
the SMARC USB0 OTG interface (client mode) when FORCE_RECOVERY# is tied
to GND during boot, enabling USB recovery.
The EC drives the FORCE_RECOVERY# signal to ground to keep the USB
recovery configuration.
The SoC USB2 is a USB 2.0 interface and does not officially support
USB recovery.

Adds support for the 'Firmware Loaded' command in the Embedded
Controller, enabling the system to switch to the default USB
configuration, as soon as the SPL is loaded.
The remaining part of the boot process can use the default configuration.

Upstream-Status: Pending
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>