toradex: integrate load protection (downstream)
This introduces an extra protection to the "load" command. When this
protection is enabled (by CONFIG_TDX_LOAD_PROTECTION) the command will
only allow the loading of files in the address range:

- [R0, R0+1G-64M) on devices with at least 1GB of SDRAM.
- [R0, R0+512M-64M) on devices with less than 1GB of SDRAM.

where R0 stands for the initial address of the SDRAM.

This is supposed to fulfill two goals:

- Prevent the running U-Boot from being overwritten.
- Prevent write access to any area that are not actual memory such as
  memory-mapped devices.

In theory, the LMB feature in U-Boot should fulfill the same goals but
tests have shown it doesn't handle all cases and seems to be even buggy.
The present implementation nonetheless adds to the LMB protections and
actually relies on data structures present only when CONFIG_LMB is
enabled.

Upstream-Status: Inappropriate [Torizon OS specific]

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>