lmb: prevent alloc overlap with reserved regions (upstream)
Add explicit check into lmb_alloc_addr() to prevent the allocated memory
range to overlap with reserved regions. Also, change the return value of
the function when the allocation is unsuccessful to prevent failures to
allocate address 0x0 from being misinterpreted as a success.

The function lmb_alloc_addr() is currently called only in the context of
the "load" command so that the present change should only impact that
particular command and no other U-Boot functionality.

Upstream-Status: Pending

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>