configs: verdin-am62p: fix R5 configuration memory addresses
The current configuration leads to overlapping .sram area and stack.
Since the overlapped area, in .sram, is not currently used, no side
effects is seen, but if that area will be used in the future, it will
lead to underfined behavior.
In particular the linker script, arch/arm/mach-omap2/u-boot-spl.lds,
uses memory from 0x43c00000 to 0x43c3e000 and the stack is set to
0x43c3a7f0.
The addresses used by am62px evm configuration avoid this by leaving a
gap between 0x43c3b000 and 0x43c4b000 and the stack is set to 0x43c4a7f0.
Fix the configuration to match addresses used by the am62px evm.

Upstream-Status: Pending
Fixes: d2aebc9910ae ("board: toradex: add verdin am62p support")
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>