arm: imx: imx9: scmi: Fix booting from USB
On i.MX95, when booting from USB, the U-Boot environment is always
assumed to be in RAM. However, this causes the boot to hang when
`CONFIG_ENV_IS_NOWHERE` is not enabled. The boot also hangs even if the
environment is present in another storage media (for example, eMMC). Fix
the issue by correctly handling the U-Boot environment's location when
booting from USB. Also, set the environment location based on the ENV
config and not solely based on the boot device type.

Upstream-Status: Submitted [https://lore.kernel.org/u-boot/20250722-v1-fix-imx95-usb-boot-v1-1-dec7f2384405@toradex.com]
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>