board: toradex: smarc-imx95: configure GPIO1 access protection in SPL
The SPL runs securely. Once U-Boot and later Linux start, we want them
to have control over the GPIOs on GPIO1, so we require non-secure,
non-privileged access at that time.

Use `spl_board_prepare_for_boot` to configure the non-secure,
non-privileged access to GPIO1 registers and interrupts after the SPL is
done and before U-Boot starts.

This used to be done in the System Manager board code (specifically in
`BRD_SM_Init`), but that runs before the SPL. The GPIO1 registers PCNS
and PCNP act like a switch, so configuring non-secure access before the
SPL runs means that the SPL will not have access. It's either secure or
non-secure access, not both at the same time.

Upstream-Status: Pending
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>