arm: dts: k3-am62-verdin: enable remoteproc support
Get the remoteproc changes that were added to linux-toradex. These
device tree nodes add support for the remoteproc framework on U-Boot.

For this patch, only the relevant remoteproc nodes were added and the
wkup_r5fss and related nodes were removed from the am62 file because for
this downstream branch there is not yet support for the WKUP R5F core.

These kernel changes have been submitted to the Linux kernel mainline
[1] and, once it is meged, later they will be synced into u-boot
mainline as OF_UPSTREAM is enabled on U-Boot mainline for the Verdin
AM62.

[1] https://lore.kernel.org/all/20250905051846.1189612-1-b-padhi@ti.com/

Upstream-Status: Inappropriate
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>