Revert "toradex: aquila-am69: dts: remove ETH_1 MCU_CLKOUT0"
This reverts commit c224a49297262922879f07a2d42f9aab2b8b3a1a.

In a SYSFW update by TI, the MCU_CLKOUT0 was changed from
25 MHz (SYSFW v11.01.08) to 50 MHz (SYSFW v11.02.02).
This change breaks the ETH_1 functionality on the Aquila AM69 V1.0,
since the ETH PHY requires a 25 MHz clock input.
To have a functional ETH_1 interface in U-Boot and Linux,
the revert overwrites the default clock value set by the new SYSFW
and configures MCU_CLKOUT0 back to 25 MHz.

Upstream-Status: Inappropriate
Signed-off-by: Franz Schnyder <franz.schnyder@toradex.com>