board: toradex: aquila-am69: enable MCU_CLKOUT0 at 25Mhz from A72 core
After updating to TI SYSFW v11.02.02, MCU_CLKOUT0 was not being enabled
correctly in the CTRLMMR_MCU_CLKOUT0_CTRL register. The present code
in `spl_board_init()` was intended to enable the MCU_CLKOUT0, but it
failed. The reason for this is that the register must be written by the
A72 core and not the R5 core.

The change also sets the CLK_SEL bit to select the 25MHz clock source
for MCU_CLKOUT0.

This provides correct functionality for the ETH_1 on Aquila AM69 V1.0
with latests SYSFW v11.02.02.

Upstream-Status: Inappropriate
Signed-off-by: Franz Schnyder <franz.schnyder@toradex.com>