drivers: usb: dwc3: gadget: add delay after gadget start
This workaround allows that Tezi works properly on the Aquila-iMX95.

Once the 'run' bit is set to the DCTL register, there is a timing issue
on the iMX95 which results in incorrect gadget initialisation and
therefore fastboot doesn't work, as it does not announce itself on the
USB bus. The delay helps to temporarily resolve the timing issue,
enabling fastboot to work. It is set to 6ms, as the issue can still
appear with a delay of 5ms. If the delay is set to 8ms, USB SDP in SPL
will fail without an additional patch series.

Link: https://lore.kernel.org/all/878qhkw8py.fsf@kernel.org/
Link: https://lore.kernel.org/all/20240412202611.3565052-1-alexander.sverdlin@siemens.com/
Upstream-Status: Inappropriate
Signed-off-by: Franz Schnyder <franz.schnyder@toradex.com>